Polybench: rename jacobi-1d-imper -> jacobi-1d
