+++ SingleSource/Benchmarks/Polybench/stencils/jacobi-1d/Makefile
@@ -1,7 +1,7 @@
 LEVEL = ../../../../..
 
 POLYBENCH_UTILS = SingleSource/Benchmarks/Polybench/utilities 
-PROG          = jacobi-1d-imper
+PROG          = jacobi-1d
 CFLAGS        += -I $(PROJ_SRC_ROOT)/$(POLYBENCH_UTILS) -DPOLYBENCH_DUMP_ARRAYS -DFP_ABSTOLERANCE=1e-5
 
 HASH_PROGRAM_OUTPUT = 1
