+++ SingleSource/Benchmarks/Polybench/stencils/jacobi-1d/jacobi-1d.c
@@ -19,7 +19,7 @@
 
 /* Include benchmark-specific header. */
 /* Default data type is double, default size is 100x10000. */
-#include "jacobi-1d-imper.h"
+#include "jacobi-1d.h"
 
 
 /* Array initialization. */
@@ -59,7 +59,7 @@ void print_array(int n,
 /* Main computational kernel. The whole function will be timed,
    including the call and return. */
 static
-void kernel_jacobi_1d_imper(int tsteps,
+void kernel_jacobi_1d(int tsteps,
 			    int n,
 			    DATA_TYPE POLYBENCH_1D(A,N,n),
 			    DATA_TYPE POLYBENCH_1D(B,N,n))
@@ -83,7 +83,7 @@ void kernel_jacobi_1d_imper(int tsteps,
 // discrepancies which cause the accuracy checks to fail.
 // In this case, the test runs with the option -ffp-contract=off
 static void
-kernel_jacobi_1d_imper_StrictFP(int tsteps,
+kernel_jacobi_1d_StrictFP(int tsteps,
                                 int n,
                                 DATA_TYPE POLYBENCH_1D(A,N,n),
                                 DATA_TYPE POLYBENCH_1D(B,N,n))
@@ -145,7 +145,7 @@ int main(int argc, char** argv)
   polybench_start_instruments;
 
   /* Run kernel. */
-  kernel_jacobi_1d_imper (tsteps, n, POLYBENCH_ARRAY(A), POLYBENCH_ARRAY(B));
+  kernel_jacobi_1d (tsteps, n, POLYBENCH_ARRAY(A), POLYBENCH_ARRAY(B));
 
   /* Stop and print timer. */
   polybench_stop_instruments;
@@ -157,7 +157,7 @@ int main(int argc, char** argv)
   polybench_prevent_dce(print_array(n, POLYBENCH_ARRAY(A)));
 #else
   init_array (n, POLYBENCH_ARRAY(A_StrictFP), POLYBENCH_ARRAY(B));
-  kernel_jacobi_1d_imper_StrictFP (tsteps, n, POLYBENCH_ARRAY(A_StrictFP),
+  kernel_jacobi_1d_StrictFP (tsteps, n, POLYBENCH_ARRAY(A_StrictFP),
                           POLYBENCH_ARRAY(B));
   if (!check_FP(n, POLYBENCH_ARRAY(A), POLYBENCH_ARRAY(A_StrictFP)))
     return 1;
