+++ SingleSource/Benchmarks/Polybench/stencils/jacobi-1d/jacobi-1d.reference_output
